Create sub-module auto_import
